Send enemies through packets instead of messages
