+++ DonjinKrawler-Common/build.gradle
@@ -1,0 +1,14 @@
+plugins {
+    id 'java'
+}
+
+version 'unspecified'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    compile "com.esotericsoftware:kryonet:2.22.0-RC1"
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+}
