+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/EnemyFactory.java
@@ -1,0 +1,5 @@
+package krawlercommon.enemies;
+
+public interface EnemyFactory {
+    public Enemy make(EnemyType enemyType);
+}
