+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/EnemyType.java
@@ -1,0 +1,7 @@
+package krawlercommon.enemies;
+
+public enum EnemyType {
+    CHICKEN,
+    SKELETON,
+    ZOMBIE;
+}
