+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/Zombie.java
@@ -1,0 +1,17 @@
+package krawlercommon.enemies;
+
+import krawlercommon.strategies.*;
+
+public abstract class Zombie extends Enemy {
+    transient EnemyStrategy[] strategies = {
+        new MoveAwayFromPlayer(),
+        new MoveRandomly(),
+        new MoveTowardPlayer(),
+        new Attack()
+    };
+
+    public Zombie() {
+        this.setInterval(2);
+        this.setStrategies(strategies);
+    }
+}
