+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/big/BigChicken.java
@@ -1,0 +1,12 @@
+package krawlercommon.enemies.big;
+
+import krawlercommon.enemies.Enemy;
+
+public class BigChicken extends Enemy {
+
+    public BigChicken() {
+        this.setName("Big-Chicken");
+        this.setDamage(15.0);
+    }
+
+}
