+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/big/BigEnemyFactory.java
@@ -1,0 +1,25 @@
+package krawlercommon.enemies.big;
+
+import krawlercommon.enemies.Enemy;
+import krawlercommon.enemies.EnemyType;
+import krawlercommon.enemies.EnemyFactory;
+
+public class BigEnemyFactory implements EnemyFactory {
+
+    public Enemy make(EnemyType type) {
+
+        if (type == null) {
+            return null;
+        }
+
+        if (type.equals(EnemyType.CHICKEN)) {
+            return new BigChicken();
+        } else if (type.equals(EnemyType.SKELETON)) {
+            return new BigSkeleton();
+        } else if (type.equals(EnemyType.ZOMBIE)) {
+            return new BigZombie();
+        }
+
+        return null;
+    }
+}
