+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/big/BigSkeleton.java
@@ -1,0 +1,11 @@
+package krawlercommon.enemies.big;
+
+import krawlercommon.enemies.Enemy;
+
+public class BigSkeleton extends Enemy {
+
+    public BigSkeleton() {
+        this.setName("Big-Skeleton");
+        this.setDamage(40.0);
+    }
+}
