+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/big/BigZombie.java
@@ -1,0 +1,12 @@
+package krawlercommon.enemies.big;
+
+import krawlercommon.enemies.Enemy;
+
+public class BigZombie extends Enemy {
+
+    public BigZombie() {
+        this.setName("Big-Zombie");
+        this.setDamage(30.0);
+    }
+
+}
