+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/small/SmallChicken.java
@@ -1,0 +1,12 @@
+package krawlercommon.enemies.small;
+
+import krawlercommon.enemies.Chicken;
+
+public class SmallChicken extends Chicken {
+
+    public SmallChicken() {
+        this.setName("Small-Chicken");
+        this.setDamage(5.0);
+    }
+
+}
