+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/small/SmallSkeleton.java
@@ -1,0 +1,11 @@
+package krawlercommon.enemies.small;
+
+import krawlercommon.enemies.Skeleton;
+
+public class SmallSkeleton extends Skeleton {
+
+    public SmallSkeleton() {
+        this.setName("Small-Skeleton");
+        this.setDamage(20.0);
+    }
+}
