+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/small/SmallZombie.java
@@ -1,0 +1,12 @@
+package krawlercommon.enemies.small;
+
+import krawlercommon.enemies.Zombie;
+
+public class SmallZombie extends Zombie {
+
+    public SmallZombie() {
+        this.setName("Small-Zombie");
+        this.setDamage(15.0);
+    }
+
+}
