+++ DonjinKrawler-Common/src/main/java/krawlercommon/packets/ConnectPacket.java
@@ -1,0 +1,5 @@
+package krawlercommon.packets;
+
+public class ConnectPacket {
+    public String name;
+}
