+++ DonjinKrawler-Common/src/main/java/krawlercommon/packets/MessagePacket.java
@@ -1,0 +1,6 @@
+package krawlercommon.packets;
+
+public class MessagePacket {
+
+    public String message;
+}
