+++ DonjinKrawler-Common/src/main/java/krawlercommon/strategies/EnemyStrategy.java
@@ -1,0 +1,11 @@
+package krawlercommon.strategies;
+
+import krawlercommon.enemies.Enemy;
+
+public interface EnemyStrategy {
+    String getStrategy();
+
+    void init(Enemy enemy);
+
+    void execute();
+}
