+++ DonjinKrawler-Common/src/main/java/krawlercommon/strategies/MoveAwayFromPlayer.java
@@ -1,0 +1,26 @@
+package krawlercommon.strategies;
+
+import krawlercommon.enemies.Enemy;
+
+public class MoveAwayFromPlayer implements EnemyStrategy {
+
+    public MoveAwayFromPlayer() {
+    }
+
+    Enemy enemy;
+
+    @Override
+    public String getStrategy() {
+        return "MoveAway";
+    }
+
+    @Override
+    public void init(Enemy enemy) {
+        this.enemy = enemy;
+    }
+
+    @Override
+    public void execute() {
+        enemy.setInfo(getStrategy());
+    }
+}
