+++ DonjinKrawler-Common/src/main/java/krawlercommon/strategies/MoveTowardPlayer.java
@@ -1,0 +1,26 @@
+package krawlercommon.strategies;
+
+import krawlercommon.enemies.Enemy;
+
+public class MoveTowardPlayer implements EnemyStrategy {
+
+    public MoveTowardPlayer() {
+    }
+
+    Enemy enemy;
+
+    @Override
+    public String getStrategy() {
+        return "MoveTowards";
+    }
+
+    @Override
+    public void init(Enemy enemy) {
+        this.enemy = enemy;
+    }
+
+    @Override
+    public void execute() {
+        enemy.setInfo(getStrategy());
+    }
+}
