+++ DonjinKrawler-Server/build.gradle
@@ -5,6 +5,7 @@ plugins {
 
 dependencies {
     compile "com.esotericsoftware:kryonet:2.22.0-RC1"
+    compile project(":DonjinKrawler-Common")
 }
 
 mainClassName = 'donjinkrawler.Server'
