--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/EnemyFactory.java
@@ -1,5 +1,0 @@
-package donjinkrawler.enemies;
-
-public interface EnemyFactory {
-    public Enemy make(EnemyType enemyType);
-}
