--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/EnemyType.java
@@ -1,7 +1,0 @@
-package donjinkrawler.enemies;
-
-public enum EnemyType {
-    CHICKEN,
-    SKELETON,
-    ZOMBIE;
-}
