--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/big/BigChicken.java
@@ -1,12 +1,0 @@
-package donjinkrawler.enemies.big;
-
-import donjinkrawler.enemies.Enemy;
-
-public class BigChicken extends Enemy {
-
-    public BigChicken() {
-        this.setName("Big-Chicken");
-        this.setDamage(15.0);
-    }
-
-}
