--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/big/BigSkeleton.java
@@ -1,11 +1,0 @@
-package donjinkrawler.enemies.big;
-
-import donjinkrawler.enemies.Enemy;
-
-public class BigSkeleton extends Enemy {
-
-    public BigSkeleton() {
-        this.setName("Big-Skeleton");
-        this.setDamage(40.0);
-    }
-}
