--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/big/BigZombie.java
@@ -1,12 +1,0 @@
-package donjinkrawler.enemies.big;
-
-import donjinkrawler.enemies.Enemy;
-
-public class BigZombie extends Enemy {
-
-    public BigZombie() {
-        this.setName("Big-Zombie");
-        this.setDamage(30.0);
-    }
-
-}
