--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/small/SmallChicken.java
@@ -1,12 +1,0 @@
-package donjinkrawler.enemies.small;
-
-import donjinkrawler.enemies.Chicken;
-
-public class SmallChicken extends Chicken {
-
-    public SmallChicken() {
-        this.setName("Small-Chicken");
-        this.setDamage(5.0);
-    }
-
-}
