--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/small/SmallSkeleton.java
@@ -1,11 +1,0 @@
-package donjinkrawler.enemies.small;
-
-import donjinkrawler.enemies.Skeleton;
-
-public class SmallSkeleton extends Skeleton {
-
-    public SmallSkeleton() {
-        this.setName("Small-Skeleton");
-        this.setDamage(20.0);
-    }
-}
