--- DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/small/SmallZombie.java
@@ -1,12 +1,0 @@
-package donjinkrawler.enemies.small;
-
-import donjinkrawler.enemies.Zombie;
-
-public class SmallZombie extends Zombie {
-
-    public SmallZombie() {
-        this.setName("Small-Zombie");
-        this.setDamage(15.0);
-    }
-
-}
