--- DonjinKrawler-Server/src/main/java/donjinkrawler/packets/ConnectPacket.java
@@ -1,5 +1,0 @@
-package donjinkrawler.packets;
-
-public class ConnectPacket {
-    public String name;
-}
