--- DonjinKrawler-Server/src/main/java/donjinkrawler/packets/MessagePacket.java
@@ -1,6 +1,0 @@
-package donjinkrawler.packets;
-
-public class MessagePacket {
-
-    public String message;
-}
