--- DonjinKrawler-Server/src/main/java/donjinkrawler/strategies/EnemyStrategy.java
@@ -1,9 +1,0 @@
-package donjinkrawler.strategies;
-
-import donjinkrawler.enemies.Enemy;
-
-public interface EnemyStrategy {
-    String getStrategy();
-    void init(Enemy enemy);
-    void execute();
-}
