--- DonjinKrawler-Server/src/main/java/donjinkrawler/strategies/MoveAwayFromPlayer.java
@@ -1,23 +1,0 @@
-package donjinkrawler.strategies;
-
-import donjinkrawler.enemies.Enemy;
-
-public class MoveAwayFromPlayer implements EnemyStrategy {
-
-    Enemy enemy;
-
-    @Override
-    public String getStrategy() {
-        return "MoveAway";
-    }
-
-    @Override
-    public void init(Enemy enemy) {
-        this.enemy = enemy;
-    }
-
-    @Override
-    public void execute() {
-        enemy.setInfo(getStrategy());
-    }
-}
