+++ gradle/gradle.properties
@@ -1,0 +1,1 @@
+org.gradle.jvmargs=-Xmx3g -XX:MaxPermSize=2048m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8 -Xss2m
