+++ settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'DonjinKrawler'
 include 'DonjinKrawler-Server'
 include 'DonjinKrawler-Client'
+include 'DonjinKrawler-Common'
+
